doc: Document the `git commit` convention that must follow
From now on, in every enterprises or companies that I have been working
for, if my title still related with developer domain specifically,
this is the rules that must have been applied for any commits in any SCM
projects that I have been participated in. If I could become a manager
/techlead, this is the first thing I have to make sure that all of my
members have to follow.